Added dedicated ValidateActivitiesInput class
